Conflicts:
	AI/Interfaces/C/CMakeLists.txt
	AI/Interfaces/Java/CMakeLists.txt
	AI/SConscript
	AI/Skirmish/KAIK/AIExport.cpp
	AI/Skirmish/NullLegacyCppAI/CMakeLists.txt
	AI/Skirmish/NullLegacyCppAI/src/AIExport.cpp
	AI/Skirmish/NullLegacyCppAI/src/TestGlobalAI.cpp
	AI/Skirmish/RAI/GTerrainMap.cpp
	AI/Skirmish/RAI/RAI.cpp
	SConstruct
	installer/make_gamedata_arch.sh
	installer/make_portable.sh
	installer/make_source_package.sh
	installer/tasclient_download.sh
	rts/System/Platform/Win/WinVersion.cpp

+++ installer/make_source_package.sh
@@ -42,13 +42,20 @@ else
 fi
 echo "Using $branch as source"
 
-# Each one of these that is set is build when running this script.
-# .tar.bz2 and .tar.gz are built with linux (LF) line endings.
-# .zip and .7z are built with windows (CRLF) line endings.
 dir="spring_${version_string}"
+
+# Each one of these that is set, is built when running this script.
+# Linux archives
+# * linux (LF) line endings
+# * removed files needed for windows installer generation only
+# * GPL compatible
 lzma="spring_${version_string}_src.tar.lzma"
 #tbz="spring_${version_string}_src.tar.bz2"
 tgz="spring_${branch}_src.tar.gz"
+
+# Windows archives
+# * windows (CRLF) line endings (bugged, see TODO below)
+# * contain everything from the GIT repository
 #zip="spring_${version_string}_src.zip"
 #seven_zip="spring_${branch}_src.7z"
 
@@ -68,12 +75,23 @@ include=" \
  $dir/CMakeLists.txt \
  $dir/VERSION \
  $dir/tools/unitsync/ \
+ $dir/tools/DemoTool/ \
  $dir/tools/DedicatedServer/"
 
 # On linux, win32 executables are useless.
-# TASClient is windows only.
 exclude_from_all=""
-linux_exclude="${exclude_from_all} $dir/installer/pkzip.exe"
+linux_exclude="${exclude_from_all}
+	${dir}/installer/include/
+	${dir}/installer/sections/
+	${dir}/installer/graphics/
+	${dir}/installer/nsis_plugins/
+	${dir}/installer/*.exe
+	${dir}/installer/*.bat
+	${dir}/installer/*.nsi
+	${dir}/installer/*.nsh
+	${dir}/installer/make_installer.pl
+	${dir}/installer/make_luaui_nsh.py
+	${dir}/installer/springlobby_download.sh"
 linux_include=""
 windows_exclude="${exclude_from_all}"
 windows_include=""
@@ -83,7 +101,8 @@ echo ""
 echo "Exporting checkout dir with LF line endings"
 git clone -n . lf/${dir} > /dev/null
 cd lf/$dir
-git checkout -b source_package_lf ${branch} > /dev/null
+git git checkout -b source_package_lf ${branch} > /dev/null
+git submodule update --init
 cd ..
 # This is used by the build systems, as they fall back to this file,
 # if the source root is not a git repository.
@@ -92,35 +111,40 @@ cd ..
 echo -n "${version_string}" > ./${dir}/VERSION
 [ -n "$linux_exclude" ] && rm -rf $linux_exclude
 [ -n "$lzma" ] && echo "Creating archive: $lzma" && \
-	tar --lzma -c -f "../$lzma" $include $linux_include
+	tar --lzma -c -f "../$lzma" $include $linux_include --exclude=.git
 [ -n "$tbz" ]  && echo "Creating archive: $tbz"  && \
-	tar -c -j -f "../$tbz" $include $linux_include
+	tar -c -j -f "../$tbz" $include $linux_include --exclude=.git
 [ -n "$tgz" ]  && echo "Creating archive: $tgz"  && \
-	tar -c -z -f "../$tgz" $include $linux_include
+	tar -c -z -f "../$tgz" $include $linux_include --exclude=.git
 cd ..
 echo "Cleaning"
 rm -rf lf
 
-### TODO: needs fixing
-# Windows line endings, .zip/.7z package
-echo ""
-echo "Exporting checkout dir with CRLF line endings"
-git clone -n . crlf/${dir} > /dev/null
-cd crlf/$dir
-git config core.autocrlf true
-git checkout -b source_package_crlf ${branch} > /dev/null
-cd ..
-# This is used by the build systems, as they fall back to this file,
-# if the source root is not a git repository.
-# They need the version to incorporate it into the library names,
-# for example: spring-0.80.5.dll
-echo -n "${version_string}" > ./${dir}/VERSION
-[ -n "$windows_exclude" ] && rm -rf $windows_exclude
-[ -n "$zip" ]       && [ -x /usr/bin/zip ] && echo "Creating archive: $zip"       && \
-	/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
-[ -n "$seven_zip" ] && [ -x /usr/bin/7z ]  && echo "Creating archive: $seven_zip" && \
-	/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include >/dev/null
-cd ..
-echo "Cleaning"
-rm -rf crlf
+if [ -n "$zip" ] || [ -n "$seven_zip" ]; then
+	### TODO: needs fixing (not really using CRLF)
+	# Windows line endings, .zip/.7z package
+	#echo 'Exporting checkout dir with CRLF line endings'
+	git clone -n . crlf/$dir > /dev/null
+	cd crlf/$dir
+	git config core.autocrlf true
+	git checkout -b source_package_crlf ${branch} > /dev/null
+	git submodule update --init
+	cd ..
+
+	# This is used by the build systems, as they fall back to this file,
+	# if the source root is not a git repository.
+	# They need the version to incorporate it into the library names,
+	# for example: spring-0.80.5.dll
+	echo -n "${version_string}" > ./${dir}/VERSION
+
+	[ -n "$windows_exclude" ] && rm -rf $windows_exclude
+	[ -n "$zip" ]       && [ -x /usr/bin/zip ] && echo "Creating archive: $zip"       && \
+		/usr/bin/zip -q -r -u -9 "../$zip" $include $windows_include
+	[ -n "$seven_zip" ] && [ -x /usr/bin/7z ]  && echo "Creating archive: $seven_zip" && \
+		/usr/bin/7z a -t7z -m0=lzma -mx=9 -mfb=64 -md=32m -ms=on "../$seven_zip" $include > /dev/null
+	cd ..
+	echo "Cleaning"
+	rm -rf crlf
+fi
+
 cd ..
